tests: Move all tests into the Python package
